Change base word to 64 bit int

+++ bitmap.h
@@ -11,7 +11,7 @@ class BitMap
 {
     // ATTRIBUTES
     private:
-        using word_t = uint8_t; //!< Type used for bit vector
+        using word_t = int64_t; //!< Type used for bit vector
         const size_t word_s = sizeof(word_t)*8; //!< Size in bits of word_t
 
         size_t _size = 0; //!< Size in bits of the bitmap
